Factor variable symbol macros out of independent packages
The use of packages, whilst keeping with lisp syntax, added an
additional layer of complexity that was not worth the return on
investment. The idea will work, but will need some more thought and
resources put into its implementation, and I don't see any real
benefit in it at this point.

From the user perspective, all that has changes is that variables are
referenced 'R style', e.g. mtcars$hp, residing in the LS-USER
package (or the value of *package* instead of 'lisp style', mtcars:hp,
with 'hp' residing in the 'mtcars' package.