Remove '$' pkg designators; improve object printing; remove guess-alist
print-object now prints a one line summary whilst describe-object will
print a full description.  This is accomplished by convention that the
short-description is the first text up to a newline.  So, if the
docstring is:

"Motor Trend Car Road Tests

This data came from the results of testing a lot of cars ..."

print-object will print as it did before:

Motor Trend Car Road Tests>

whilst describe-object will include in its documentation the entire
description.

This commit also removes some additional vestiges of the '$' package
designator and removes guess-alist? in favour of alexandria+:alistp